Tasks
submitted, a return request is created and user is redirected to ‘Add Item’ screen
4. Create ‘Add Item’ screen that prompts the user to enter the following data: (For missing info use
data provided in above APIs)
a. NDC
b. Description
c. Manufacturer
d. Full Quantity
e. Partial Quantity
f. Expiration Date
g. Lot Number
*Once item is added, show a successful message to the user, clear the form to prompt the user to add a new item
*Create a button that redirects the user to the ‘Items’ screen when he’s done
5. Create ‘Items’ screen that lists all added items of a certain return request with the same information entered when adding an item
a. Add deletion functionality to an item in this screen
b. Add updating functionality to an item in this screen (You can prompt the user to update
only the description of an item, send the full object with only the description changed)